vm: move changes of old state manager into new one
